LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/generic-family-keywords-001-expected.txt

00000
00000

PASS @font-face matching for quoted and unquoted serif
PASS @font-face matching for quoted and unquoted sans-serif
PASS @font-face matching for quoted and unquoted cursive
PASS @font-face matching for quoted and unquoted fantasy
PASS @font-face matching for quoted and unquoted monospace
FAIL @font-face matching for quoted and unquoted system-ui assert_equals: quoted system-ui matches  @font-face rule expected 50 but got 33
FAIL @font-face matching for quoted and unquoted math assert_equals: quoted math matches  @font-face rule expected 50 but got 25
PASS @font-face matching for quoted and unquoted generic(fangsong)
PASS @font-face matching for quoted and unquoted generic(kai)
PASS @font-face matching for quoted and unquoted generic(khmer-mul)
PASS @font-face matching for quoted and unquoted generic(nastaliq)
FAIL @font-face matching for quoted and unquoted ui-serif assert_equals: quoted ui-serif matches  @font-face rule expected 50 but got 32
FAIL @font-face matching for quoted and unquoted ui-sans-serif assert_equals: quoted ui-sans-serif matches  @font-face rule expected 50 but got 33
FAIL @font-face matching for quoted and unquoted ui-monospace assert_equals: quoted ui-monospace matches  @font-face rule expected 50 but got 31
FAIL @font-face matching for quoted and unquoted ui-rounded assert_equals: quoted ui-rounded matches  @font-face rule expected 50 but got 34
